swasth-scaler/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.30.0
httpx==0.24.1
python-multipart==0.0.9
twilio==9.0.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
